requirements.txt
# Основные зависимости
python-telegram-bot==20.7
python-dotenv==1.0.0

# API сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# База данных
sqlite3

# Дополнительные
aiofiles==23.2.1
python-multipart==0.0.6